fix: make fields in PlayerInfo::Action public (#32)
This PR makes the fields in the PlayerInfo packet public.

I encountered this issue when writing my own bot which needs a list of UUID > String of all currently logged-in players, and while writing my handler for `Action::RemovePlayer` I found out that the `uuid` field is private. This PR fixes that.